feat: Add support for lexicon replacement in SSMLConverter
This commit adds support for lexicon replacement in the SSMLConverter class. The `loadLexicon` method is implemented to load a lexicon for SSML conversion. The lexicon is used to replace specific words or phrases with SSML tags, such as `<sub>` for aliases and `<phoneme>` for phonemes. The lexicon is case-sensitive, and multiple words can be replaced at once.

The code changes include:
- Adding the `loadLexicon` method to the `SSMLConverter` class
- Implementing lexicon replacement in the `addText` method

This feature enhances the functionality of the SSMLConverter by allowing users to easily add lexicon replacements to their SSML documents.